Address PR review: co-locate protocols, add CatalogWriteConfig, remove dbt-core import
- Move CatalogV2 Protocol into _integration.py alongside CatalogIntegrationConfig;
  delete _protocols.py
- Add CatalogWriteConfig concrete dataclass to dbt-adapters so bridge_v2_catalog
  has no runtime dependency on dbt-core
- bridge_v2_catalog returns CatalogWriteConfig (satisfies CatalogIntegrationConfig)
- Simplify base adapter tests: no longer need sys.modules mock
