+++ dbt-adapters/src/dbt/adapters/catalogs/__init__.py
@@ -9,7 +9,8 @@ from dbt.adapters.catalogs._integration import (
     CatalogIntegration,
     CatalogIntegrationConfig,
     CatalogRelation,
+    CatalogV2,
+    CatalogWriteConfig,
 )
 
 from dbt.adapters.catalogs._constants import CATALOG_INTEGRATION_MODEL_CONFIG_NAME
-from dbt.adapters.catalogs._protocols import CatalogV2
